back/requirements.txt
Flask-Login==0.6.3
Flask==3.0.3

python-dotenv==1.0.1
websockets==13.0.1

psycopg2-binary==2.9.9